imessage-bots/src/bots/lover-bot/pyproject.toml
[tool.poetry]
name = "lover-bot"
version = "2.0.0"
description = "An AI girlfriend bot that provides context-aware romantic messaging"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "lover_bot", from = "."}]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.100.0"
uvicorn = "^0.20.0"
pydantic = "^2.0.0"
requests = "^2.28.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
imessage-bot-framework = {path = "../../../../", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 